fix(macos): bind background notices to recorded run outcomes
